Allow >MAX_INT values for policy id

+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -85,12 +85,12 @@ public class ParserUtil {
     /**
      * Ensure non-null and is numeric
      */
-    public static int parsePolicyNumber(String number) throws ParseException {
+    public static long parsePolicyNumber(String number) throws ParseException {
         requireNonNull(number);
         if (!number.matches(Policy.NUMBER_VALIDATION_REGEX)) {
             throw new ParseException(INTEGER_MESSAGE_CONSTRAINTS);
         }
-        return Integer.valueOf(number);
+        return Long.valueOf(number);
     }
 
     /**
